+++ src/process_manager.c
@@ -50,6 +50,16 @@ static bool process_is_observable(struct process *process)
         return false;
     }
 
+    if (string_equals(process->name, "loginwindow")) {
+        debug("%s: %s (%d) is blacklisted! ignoring..\n", __FUNCTION__, process->name, process->pid);
+        return false;
+    }
+
+    if (string_equals(process->name, "ScreenSaverEngine")) {
+        debug("%s: %s (%d) is blacklisted! ignoring..\n", __FUNCTION__, process->name, process->pid);
+        return false;
+    }
+
     return true;
 }